Writing, closing sync device drivers

+++ kernel/src/devices/mod.rs
@@ -6,8 +6,8 @@ use crate::task::id::TaskID;
 
 pub trait SyncDriver {
     fn open(&self) -> Result<u32, ()>;
-    fn read(&self, index: u32, buffer: &mut [u8]) -> Result<u32, ()>;
-    fn write(&self, index: u32, buffer: &[u8]) -> Result<u32, ()>;
+    fn read(&self, index: u32, buffer: &mut [u8]) -> Result<usize, ()>;
+    fn write(&self, index: u32, buffer: &[u8]) -> Result<usize, ()>;
     fn close(&self, index: u32) -> Result<(), ()>;
 }
